domain: use double instead of long for value

+++ src/TTX.Domain/Models/PortfolioSnapshot.cs
@@ -4,7 +4,7 @@ namespace TTX.Domain.Models;
 
 public class PortfolioSnapshot
 {
-    public required long Value { get; init; }
+    public required double Value { get; init; }
     public required ModelId PlayerId { get; init; }
     public DateTimeOffset Time { get; init; } = DateTime.UtcNow;
